requirements.txt
opencv-python
numpy
nltk
spacy
tkinter
PIL
mysql-connector-python
scipy
gunicorn
flask
flask-cors
stanfordcorenlp

